requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
